requirements.txt
telethon
tgcrypto
requests
python-dotenv
curl_cffi
fake_useragent
fastapi
uvicorn
jinja2
